+++ src/commands/player/Friend.ts
@@ -213,7 +213,8 @@ export class Command extends BaseCommand {
         const sent = await msg.channel.send(embed);
 
         if (pageLimit > 1) await sent.react("◀️");
-        await sent.react("754832389620105276");
+        if (this.permissions.MANAGE_MESSAGES)
+          await sent.react("754832389620105276");
         if (pageLimit > 1) await sent.react("▶️");
 
         let filter;
@@ -247,7 +248,10 @@ export class Command extends BaseCommand {
                   await this.parseFriends(newFriends, executor, msg.client)
                 )
             );
-          } else if (r.emoji.name === "delete") {
+          } else if (
+            r.emoji.name === "delete" &&
+            this.permissions.MANAGE_MESSAGES
+          ) {
             (<TextChannel>msg.channel).bulkDelete([msg, sent]);
           } else if (r.emoji.name === "▶️" && page !== pageLimit) {
             page++;
@@ -269,7 +273,8 @@ export class Command extends BaseCommand {
           r.users.remove(msg.author);
         });
         collector.on("end", async () => {
-          if (!sent.deleted) sent.reactions.removeAll();
+          if (!sent.deleted && this.permissions.MANAGE_MESSAGES)
+            sent.reactions.removeAll();
         });
         return;
       }
